backend2/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
pandas==2.0.3
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.0
